Changed how a button component is made/style
Previously, a button component was a template and it was included in another template.
However, that solution feels really scuffed and not flexible.
Therefore, a more of styling approach would solve the problem.
This approach focuses on how to make an element look like a button instead of worrying about using the right tag.
As a result, when developing a page, there is no need to worry about the type of the button since it can be defined as a button, anchor or input.
Adding additional classes to fit the current page is also possible because the button is not bound to a pre-defined template.